Allow the version of Node.js to download to specified
Instead of hardcoding the Node.js version to download when building apm,
it can now be overridden which right now allows us to continue migrating
Atom towards running on Electron, yet not break Atom running on atom-shell.
In the future this should also ease the process of testing newer releases
of Electron.

This does not currently support io.js, but that should be resolved.