Ensure --harmony_collections flag is set for Node.js v0.10.35
This is essentially a temporary solution to avoid breaking Atom running
atom-shell. Note that this only handles the simple scenario of not breaking
the apm targeting Node.js v0.10.35, which is the default for apm bundled
with Atom running on atom-shell. In other words, if someone is using a
Node.js version other than 0.10.35 and less than 0.12.x, they will not
have the required --harmony_collections flag set. I don't think it's worth
worrying about this however.