Wait 10s for the tag to be found on `apm publish`
I think there are situations where 5s is not enough, especially when Atom has to
authenticate with GH. https://discuss.atom.io/t/cant-publish-package/4890/16?u=joshcheek

This will still work as soon as the tag is found, but will longer if it hasn't
been found yet. This should be a rare use case, so I don't expect it to add much overhead,
but when it goes wrong, it's difficult to figure out what to do (perhaps a --retry) flag,
so increasing the time limit will save a lot of frustration in that case (assuming the
it just took longer than the timeout).

Notes
-----

1. While I suspect that the timeout duration is the cause of the issue linked above,
   I do not know this for a fact (I haven't reproduced it as I don't know the cause)
2. No tests exercise this code. On `./bin/npm test`, all tests pass with this change
   but no tests fail when I add a throw statement on the line above. Seems benign enough,
   but maybe look it over more seriously for that reason