Add flag to not list packages' versions
This commit adds a new command line option to the `list` command that
allows including, or skipping the version information when listing
packages. This is done with a `--versions` flag that is, by default,
always true. To get the false behaviour, it's as simple as prepending
it with `no`, such as `--no-versions`.

Another noteworthy thing with this commit is the added identation on
line 72. Previously, when running `apm list` either on the default, or
on bare, it would print an extra newline for every new type of
package, so say between user and bundled packages there would be a
newline. This is useful for the pretty default version of list,
however, on the bare format, since there's no info listing what type
of package each one is, unlike with the default format, I consider it
unhelpful. Both for external script usage, and because it complicates
tests with newlines that don't separate anything meaningful.

So I've kept that newline for the default format, but removed it for
the bare format.