src: Do not set the `--msvs_version=` flag
In modern node-gyp, this flag only serves to limit
node-gyp to using a specific version of Visual Studio.

It can't make node-gyp find a version it wouldn't have found.

Our checks are much less sophisticated at detecting usable
installations of Visual Studio. We should let node-gyp
use the install of Visual Studio/Build Tools it thinks is best.

It will pick any usable Visual Studio install, preferring the newest
if multiple are installed and usable. It generally will not pick one
mistakenly that it actually can't use.

Given that the stakes are "we might prevent a valid installation
of Visual Studio from being used," and there is no benefit to be had,
this flag should not be set.