Pass --verbose flag to npm if set for apm install
`apm install` can now forward the `--verbose` flag to `npm`,
forming the `npm install --verbose` command under the hood.

This prints a lot of valuable information, which can be used
to debug what happens when users try to install packages
and dependencies.

(Prints the versions of node-gyp, Python, Visual Studio and Electron
that are being used, among other things.)