:checkered-flag: Add `getResourcePath()` fallback for Windows
Adds fallback behavior for `getResourcePath()` on Windows
(roughly equivalent to the behavior on Linux and macOS).

This should allow apm to find an installation of Atom on the disk,
even when `apm` is not bundled with the Atom install it's looking for.
For example: when working in the apm git repo.